Fix an incorrect type cast in an assignment in vdbe.c.
FossilOrigin-Name: 320b5295da57023399c7b9eed82eae40cf619803

+++ src/vdbe.c
@@ -3515,7 +3515,7 @@ case OP_SeekGt: {       /* jump, in3 */
       **     r.flags = 0;
       **   }
       */
-      r.flags = (u16)(UNPACKED_INCRKEY * (1 & (oc - OP_SeekLt)));
+      r.flags = (u8)(UNPACKED_INCRKEY * (1 & (oc - OP_SeekLt)));
       assert( oc!=OP_SeekGt || r.flags==UNPACKED_INCRKEY );
       assert( oc!=OP_SeekLe || r.flags==UNPACKED_INCRKEY );
       assert( oc!=OP_SeekGe || r.flags==0 );
